Revert "Revert "adb: detect some spin loops and abort.""
This reverts commit 9fbc204f615c4a1f4bd4663e73ef01e2d8284698 [formerly 04b9ca8c31e789db68858641a21424f99c02578c].

The original incarnation of this patch was falsely triggering when our
poll would block with no activity happening for an extended amount of
time. When the poll returned, we would immediately flag that as a spin
loop. Solve this by tracking the last time we looped, to detect this.

While we're at it, switch from CLOCK_MONOTONIC to CLOCK_BOOTTIME, for
the same reason.

Change-Id: Ibcdf69d93f7b6012142cafd72066f39494c1f84b
Test: ./test_device.py

Former-commit-id: 39c1f4bca5cf2f1c04286876c15b9e689df0e193